backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
python-dotenv>=1.0.0
qdrant-client>=1.12.0
sentence-transformers>=3.0.0
pydantic>=2.9.0
beautifulsoup4>=4.12.0
requests>=2.32.0
selenium>=4.25.0
undetected-chromedriver>=3.5.0
webdriver-manager>=4.0.0
urllib3>=2.2.0
tqdm
qdrant-client==1.17.1
prometheus-fastapi-instrumentator>=7.0.0
